Move XMR encoder/decoder logic inside it's own folder
